update sdk 5.4.1.1

+++ QCloudCSharpSDK/COSXML/Transfer/TransferManager.cs
@@ -11,22 +11,22 @@ namespace COSXML.Transfer
 {
     public sealed class TransferConfig
     {
-        internal int divisionForCopy = 5242880; // 5M
+        internal long divisionForCopy = 5242880; // 5M
 
-        internal int sliceSizeForCopy = 2097152; // 2M
+        internal long sliceSizeForCopy = 2097152; // 2M
 
-        internal int divisionForUpload = 5242880; // 5M
+        internal long divisionForUpload = 5242880; // 5M
 
-        internal int sliceSizeForUpload = 1048576; // 1M
+        internal long sliceSizeForUpload = 1048576; // 1M
 
 
-        public int DdivisionForCopy { get { return divisionForCopy; } set { divisionForCopy = value; } }
+        public long DdivisionForCopy { get { return divisionForCopy; } set { divisionForCopy = value; } }
 
-        public int DivisionForUpload { get { return divisionForUpload; } set { divisionForUpload = value; } }
+        public long DivisionForUpload { get { return divisionForUpload; } set { divisionForUpload = value; } }
 
-        public int SliceSizeForCopy { get { return sliceSizeForCopy; } set { sliceSizeForCopy = value; } }
+        public long SliceSizeForCopy { get { return sliceSizeForCopy; } set { sliceSizeForCopy = value; } }
 
-        public int SliceSizeForUpload { get { return sliceSizeForUpload; } set { sliceSizeForUpload = value; } }
+        public long SliceSizeForUpload { get { return sliceSizeForUpload; } set { sliceSizeForUpload = value; } }
     }
 
     public sealed class TransferManager
